Update examples to use new methods.
